+++ include/mujoco/mjspecmacro.h
@@ -114,6 +114,7 @@
     X   ( mjtByte,         mocap,            1 ) \
     X   ( double,          gravcomp,         1 ) \
     X   ( mjtSleepPolicy,  sleep,            1 ) \
+    X   ( mjtByte,         simple,           1 ) \
     X   ( mjDoubleVec*,    userdata,         1 ) \
     X   ( mjtByte,         explicitinertial, 1 ) \
     X   ( mjsPlugin,       plugin,           1 ) \